requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
mysql-connector-python==9.2.0
tensorflow==2.19.0
pillow==11.1.0
numpy==2.1.3
python-multipart==0.0.20
pandas==2.2.3